+++ CHANGELOG.md
@@ -4,12 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.2] - 2019-03-16
 ### Added
 - Information about installing from release
+- Added linting guideline to readme for contributors
+- Remove old containers on stacks rm
+
+### Fixed
+- Logs output now includes stderr
 
 ### Changed
 - Changed the order of the readme file
+- Lint with gofmt
 
 ## [0.0.1] - 2019-03-03
 ### Changed